add new file PostTableViewCell, add new cell and tableView
