+++ Navigation/Assets.xcassets/1.dataset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "data" : [
+    {
+      "filename" : "Unknown",
+      "idiom" : "universal"
+    }
+  ],
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
